Do not purge test SDK directory when shutting down

+++ server/src/main/java/com/defold/extender/services/DefoldSdkService.java
@@ -25,12 +25,15 @@ import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Collections;
 import java.util.Comparator;
+import java.util.List;
 
 @Service
 public class DefoldSdkService {
     private static final Logger LOGGER = LoggerFactory.getLogger(DefoldSdkService.class);
     private static final String REMOTE_SDK_URL_PATTERN = "http://d.defold.com/archive/%s/engine/defoldsdk.zip";
+    private static final List<String> KEEP_SDK_DIRECTORIES = Collections.singletonList("a");
 
     private final Path baseSdkDirectory;
     private final File dynamoHome;
@@ -105,10 +108,16 @@ public class DefoldSdkService {
     }
 
     private void deleteCachedSdk(Path path) {
+        final String directoryToDelete = path.toAbsolutePath().toString();
+
+        if (KEEP_SDK_DIRECTORIES.contains(directoryToDelete)) {
+            return;
+        }
+
         try {
             FileUtils.deleteDirectory(path.toFile());
         } catch (IOException e) {
-            LOGGER.error("Failed to delete cached SDK at " + path.toAbsolutePath().toString(), e);
+            LOGGER.error("Failed to delete cached SDK at " + directoryToDelete, e);
         }
     }
